+++ app/src/main/java/com/hucet/clean/gallery/model/DialogRadioItem.kt
@@ -3,4 +3,4 @@ package com.hucet.clean.gallery.model
 /**
  * Created by taesu on 2017-11-17.
  */
-data class DialogRadioItem(val index: Int, val title: String, val isCheck: Boolean, val bitAtt: Int)
+data class DialogRadioItem(val index: Int, val title: String, val isCheck: Boolean, val bitAtt: Long)
